Video Handler Code Routes Cleanup
